feat: register controller
